+++ packages/orq-rc/src/models/operations/getallprompts.ts
@@ -48,6 +48,7 @@ export const GetAllPromptsModelType = {
   Rerank: "rerank",
   Ocr: "ocr",
   Moderation: "moderation",
+  Classify: "classify",
   Vision: "vision",
 } as const;
 /**
@@ -448,6 +449,7 @@ export const GetAllPromptsProvider = {
   Reson8: "reson8",
   Meta: "meta",
   Greenpt: "greenpt",
+  Typesafe: "typesafe",
   Slack: "slack",
   Orq: "orq",
 } as const;