+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -176,7 +176,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-09-18T06:13:18.898Z",
+    "2026-09-18T12:12:53.053Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {