+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -3949,7 +3949,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01M2SJBWVTJX43T64A6SJ52526"),
+  id: z.string().default("01M2T6Y9KKCT8JCEDTX0W9DKS9"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>