+++ packages/orq-rc/RELEASES.md
@@ -12479,3 +12479,13 @@ Based on:
 - [typescript v4.16.0-rc.3] packages/orq-rc
 ### Releases
 - [NPM v4.16.0-rc.3] https://www.npmjs.com/package/@orq-ai/node/v/4.16.0-rc.3 - packages/orq-rc
+
+## 2026-09-18 12:15:25
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.797.0 (2.937.18) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.16.0-rc.4] packages/orq-rc
+### Releases
+- [NPM v4.16.0-rc.4] https://www.npmjs.com/package/@orq-ai/node/v/4.16.0-rc.4 - packages/orq-rc